[development/directx_init_drawing] change u32 to s32

+++ source/1_engine/rex_engine/include/rex_engine/types.h
@@ -45,9 +45,9 @@ using a_s32  = rsl::atomic<s32>;
   #define rex_atomic_load(a) a.load() // NOLINT(readability-identifier-naming)
 #endif
 
-#define REX_INVALID_INDEX ((u32)-1)
+#define REX_INVALID_INDEX -1
 
-inline bool is_valid(u32 handle)
+inline bool is_valid(s32 handle)
 {
     return handle != REX_INVALID_INDEX;
 }
